fix: resolve 7 bugs from the adversarial multi-agent review
A 5-dimension review (each finding verified by an independent skeptic) confirmed
7 real bugs; all are fixed and re-verified end-to-end (17/17 bridge checks):

IpcClient (transport):
- Ordering race: per-frame detached Tasks could reorder inbound frames and let a
  close be observed before earlier frames. Frames now flow through a single
  ordered consumer that dispatches each before reading the next.
- fd-reuse race: disconnect now calls shutdown(SHUT_RDWR) before close() so the
  blocked recv is reliably woken and a recycled fd can't be read by a stale read.
- events stream: now bounded (bufferingNewest) and finished on teardown, so a
  consumer's `for await` completes and the buffer can't grow without bound.

CoreConnection:
- connect() re-entrancy guard was ineffective (flag set after the first await);
  a synchronous in-flight flag now prevents overlapping connects leaking a client.

Core:
- config.set applied changes to shared state before persisting; on save failure
  RAM and settings.json diverged. Now mutates a clone, saves, then commits.
- accept loop now backs off on persistent accept() errors instead of busy-spinning.

MessagePack:
- float32 (0xCA) was reinterpreted as a 64-bit double pattern (garbage); now
  widened from a 32-bit Float. Added a regression test.

+++ app/Sources/ClaudeStudioKit/MessagePack.swift
@@ -319,7 +319,8 @@ public enum MessagePack {
             case 0xC6:
                 return .binary(try readData(count: Int(try readUInt32())))
             case 0xCA:
-                return .double(Double(bitPattern: UInt64(try readUInt32())))
+                // float32: load as a 32-bit Float bit pattern, then widen.
+                return .double(Double(Float(bitPattern: try readUInt32())))
             case 0xCB:
                 return .double(Double(bitPattern: try readUInt64()))
             case 0xCC:
